+++ src/main/java/com/iexec/worker/sgx/SgxService.java
@@ -17,6 +17,7 @@
 package com.iexec.worker.sgx;
 
 import com.github.dockerjava.api.model.Device;
+import com.github.dockerjava.api.model.HostConfig;
 import com.iexec.commons.containers.DockerRunFinalStatus;
 import com.iexec.commons.containers.DockerRunRequest;
 import com.iexec.commons.containers.DockerRunResponse;
@@ -24,6 +25,7 @@ import com.iexec.commons.containers.SgxDriverMode;
 import com.iexec.commons.containers.SgxUtils;
 import com.iexec.worker.config.WorkerConfigurationService;
 import com.iexec.worker.docker.DockerService;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
@@ -43,8 +45,9 @@ public class SgxService {
     private final WorkerConfigurationService workerConfigService;
     private final ApplicationContext context;
     private final DockerService dockerService;
+    @Getter
     private final SgxDriverMode sgxDriverMode;
-
+    @Getter
     private boolean sgxEnabled;
 
     public SgxService(
@@ -75,12 +78,10 @@ public class SgxService {
         }
     }
 
-    public SgxDriverMode getSgxDriverMode() {
-        return sgxDriverMode;
-    }
-
-    public boolean isSgxEnabled() {
-        return sgxEnabled;
+    public List<Device> getSgxDevices() {
+        return Arrays.stream(sgxDriverMode.getDevices())
+                .map(Device::parse)
+                .collect(Collectors.toList());
     }
 
     private boolean isSgxSupported(SgxDriverMode sgxDriverMode) {
@@ -109,7 +110,7 @@ public class SgxService {
 
         final String[] devices = sgxDriverMode.getDevices();
         // Check all required devices exists
-        final String cmd = String.format("/bin/sh -c '%s; echo $?;'",
+        final String cmd = String.format("/bin/sh -c '%s'",
                 Arrays.stream(devices)
                         .map(devicePath -> "test -e \"" + devicePath + "\"")
                         .collect(Collectors.joining(" && ")));
@@ -123,12 +124,14 @@ public class SgxService {
                 .map(Device::parse)
                 .collect(Collectors.toList());
 
+        HostConfig hostConfig = HostConfig.newHostConfig()
+                .withDevices(devicesBind);
         DockerRunRequest dockerRunRequest = DockerRunRequest.builder()
+                .hostConfig(hostConfig)
                 .containerName(containerName)
                 .imageUri(alpineLatest)
                 .cmd(cmd)
                 .maxExecutionTime(60000) // 1 min
-                .devices(devicesBind)
                 .build();
 
 
@@ -139,7 +142,7 @@ public class SgxService {
         }
 
         // Check test returned a 0 exit code.
-        String testResult = dockerRunResponse.getStdout().trim();
-        return Integer.parseInt(testResult) == 0;
+        log.info("docker container output {}", dockerRunResponse.getStdout().trim());
+        return true;
     }
 }